+++ src/Transifex.php
@@ -165,15 +165,16 @@ class Transifex
 	/**
 	 * Get an option from the Transifex instance.
 	 *
-	 * @param   string  $key  The name of the option to get.
+	 * @param   string  $key      The name of the option to get.
+	 * @param   mixed   $default  The default value if the option is not set.
 	 *
 	 * @return  mixed  The option value.
 	 *
 	 * @since   1.0
 	 */
-	public function getOption($key)
+	public function getOption($key, $default = null)
 	{
-		return isset($this->options[$key]) ? $this->options[$key] : null;
+		return isset($this->options[$key]) ? $this->options[$key] : $default;
 	}
 
 	/**